Dodano walidacje

+++ src/main/java/pl/zagorski/services/DeliveryServiceImpl.java
@@ -7,6 +7,7 @@ import pl.zagorski.domain.Delivery;
 import pl.zagorski.domain.Employee;
 import pl.zagorski.domain.PurchaseOrder;
 import pl.zagorski.domain.Warehouse;
+import pl.zagorski.exceptions.ExceptionSample;
 import pl.zagorski.repositories.DeliveryDao;
 import pl.zagorski.repositories.EmployeeRepositoryImpl;
 import pl.zagorski.repositories.PurchaseOrderRepositoryImpl;
@@ -15,6 +16,7 @@ import pl.zagorski.repositories.StatusRepositoryImpl;
 import java.sql.Date;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -76,34 +78,34 @@ public class DeliveryServiceImpl implements DeliveryImpl {
 
     @Override
     @Transactional
-    public void save(String expirationDate, int idOrderAddDelivery, String userLogin) {
-        Delivery delivery = new Delivery();
-
-        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-        java.util.Date parsed = null;
-        try {
-            parsed = format.parse(expirationDate);
-        } catch (ParseException e) {
-            e.printStackTrace();
+    public void save(String expirationDate, int idOrderAddDelivery, String userLogin) throws ExceptionSample {
+        //lek moze miec maksymalnie 3 lata waznosci
+        if(purchaseOrderRepository.findOne(idOrderAddDelivery) != null &&
+                ((Integer.parseInt(expirationDate.substring(0,4))) <= LocalDate.now().getYear() + 3) &&
+                (Integer.parseInt(expirationDate.substring(0,4))) >= LocalDate.now().getYear()){
+            Delivery delivery = new Delivery();
+            SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
+            java.util.Date parsed = null;
+            try {
+                parsed = format.parse(expirationDate);
+            } catch (ParseException e) {
+                e.printStackTrace();
+            }
+            java.sql.Date sql = new java.sql.Date(parsed.getTime());
+            delivery.setExpiration_date(sql);
+            long time = System.currentTimeMillis();
+            delivery.setDelivery_date(new Date(time));
+            Employee employee = employeeRepository.getEmployeeByLogin(userLogin).get();
+            delivery.setEmployee(employee);
+            PurchaseOrder purchaseOrder = purchaseOrderRepository.findOne(idOrderAddDelivery);
+            purchaseOrder.setStatus(statusRepository.getStatusByName(DeliveryServiceImpl.DELIVERED_STATUS));
+            delivery.setOrder(purchaseOrder);
+            purchaseOrderRepository.edit(purchaseOrder);
+            deliveryDao.save(delivery);
+            warehouseService.save(purchaseOrder.getAmount(),new Date(time),sql,employee,purchaseOrder);
+        }else{
+            throw new ExceptionSample("Dostarczone dane nie mogą zostać uznane za prawidłowe.");
         }
-        java.sql.Date sql = new java.sql.Date(parsed.getTime());
-        delivery.setExpiration_date(sql);
-
-        long time = System.currentTimeMillis();
-        delivery.setDelivery_date(new Date(time));
-
-        Employee employee = employeeRepository.getEmployeeByLogin(userLogin).get();
-        delivery.setEmployee(employee);
-
-        PurchaseOrder purchaseOrder = purchaseOrderRepository.findOne(idOrderAddDelivery);
-        purchaseOrder.setStatus(statusRepository.getStatusByName(DeliveryServiceImpl.DELIVERED_STATUS));
-        delivery.setOrder(purchaseOrder);
-
-        purchaseOrderRepository.edit(purchaseOrder);
-        deliveryDao.save(delivery);
-
-        warehouseService.save(purchaseOrder.getAmount(),new Date(time),sql,employee,purchaseOrder);
-
     }
 
     @Override
@@ -114,12 +116,17 @@ public class DeliveryServiceImpl implements DeliveryImpl {
 
     @Override
     @Transactional
-    public void delete(int deliveryDelete) {
-        Delivery delivery = findOne(deliveryDelete);
-        int purchaseOrderId = delivery.getOrder().getId();
-        warehouseService.delete(purchaseOrderId);
-        deliveryDao.delete(delivery);
-        purchaseOrderService.delete(purchaseOrderId);
+    public void delete(int deliveryDelete) throws ExceptionSample {
+        if(findOne(deliveryDelete) != null){
+            Delivery delivery = findOne(deliveryDelete);
+            int purchaseOrderId = delivery.getOrder().getId();
+            warehouseService.delete(purchaseOrderId);
+            deliveryDao.delete(delivery);
+            purchaseOrderService.delete(purchaseOrderId);
+        }else{
+            throw new  ExceptionSample("Dostarczone dane nie mogą zostać uznane za prawidłowe.");
+        }
+
     }
 
     @Override
